+++ czkawka_core/src/empty_files.rs
@@ -63,6 +63,7 @@ impl EmptyFiles {
             .allowed_extensions(self.common_data.allowed_extensions.clone())
             .excluded_items(self.common_data.excluded_items.clone())
             .recursive_search(self.common_data.recursive_search)
+            .tool_type(self.common_data.tool_type)
             .build()
             .run();