+++ czkawka_core/src/same_music.rs
@@ -203,6 +203,7 @@ impl SameMusic {
             .allowed_extensions(self.common_data.allowed_extensions.clone())
             .excluded_items(self.common_data.excluded_items.clone())
             .recursive_search(self.common_data.recursive_search)
+            .tool_type(self.common_data.tool_type)
             .max_stage(max_stage)
             .build()
             .run();