+++ krokiet/src/common.rs
@@ -23,3 +23,7 @@ pub fn create_string_standard_list_view_from_pathbuf(items: &[PathBuf]) -> Model
         .collect::<Vec<_>>();
     ModelRc::new(VecModel::from(new_folders_standard_list_view))
 }
+
+pub fn create_vec_model_from_vec_string(items: Vec<String>) -> VecModel<SharedString> {
+    VecModel::from(items.into_iter().map(SharedString::from).collect::<Vec<_>>())
+}